Rename traits and their files
