+++ contracts/cw-staking/src/handlers/execute.rs
@@ -1,7 +1,7 @@
 use crate::contract::{CwStakingAdapter, CwStakingResult};
 use crate::msg::{CwStakingAction, CwStakingExecuteMsg, ProviderName, IBC_STAKING_PROVIDER_ID};
 use crate::providers::resolver::{self, is_over_ibc};
-use crate::LocalStakingCommand;
+use crate::StakingAdapter;
 use abstract_sdk::core::ibc_client::CallbackInfo;
 use abstract_sdk::feature_objects::AnsHost;
 use abstract_sdk::features::{AbstractNameService, AbstractResponse};
