+++ contracts/cw-staking/src/traits/mod.rs
@@ -1,3 +1,3 @@
-pub mod cw_staking_adapter;
+pub mod command;
 pub(crate) mod identify;
-pub mod local_cw_staking;
+pub mod adapter;
